Ch tests 161611030

+++ api/views/products_views.py
@@ -1,17 +1,18 @@
 from flask import jsonify, request, abort, Blueprint
 from api.models.products import Products
 from flask_jwt_extended import jwt_required
+from api.views.user_views import user_check
 
 prodbp = Blueprint('prod', __name__)
 
 product_cls = Products()
+
 
 @prodbp.route('admin/products', methods=['POST'])
 @jwt_required
 def add_product():
     data = request.json
 
-    category = data['category']
     product_name = data['product_name']
     product_specs = data['product_specs']
     product_stock = int(data['product_stock'])
@@ -19,42 +20,79 @@ def add_product():
 
     if not product_name or not product_price or not product_stock:
         abort(400)
-    
-    product_cls.add_product(category = category, product_name = product_name,
-                            product_specs = product_specs,
-                            product_price = product_price, product_stock = product_stock)
+
+    if user_check() is False:
+        return jsonify({"Alert": "You're not Authorized to perform action"}), 401
+
+    product_cls.add_product(product_name=product_name,
+                            product_specs=product_specs,
+                            product_price=product_price,
+                            product_stock=product_stock)
 
     return jsonify({"Success": "The product has been added"}), 201
 
 
 @prodbp.route('admin/products', methods=['GET'])
 def view_all_products():
-    return jsonify({"Products":product_cls.get_all_products()})
+    return jsonify({"Products": product_cls.get_all_products()}), 200
 
 
 @prodbp.route('admin/products/<int:product_id>', methods=['GET'])
 def view_one_product(product_id):
-    return jsonify({"Product":product_cls.get_one_product_by_id(product_id)}), 200
+    product = product_cls.get_one_product_by_id(product_id)
+    if not product:
+        return jsonify({"Not Found":
+                        "The product with ID '{0}' was not found".format(product_id)}), 404
 
-# @prodbp.route('admin/products/<int:product_id>', methods=['GET'])
-# @jwt_required
-# def view_product_category(product_id):
-#     return jsonify({"Product":product_cls.get_one_product_by_id(product_id)}), 200
+    return jsonify({"Product": product}), 200
+
 
 @prodbp.route('admin/products/<product_id>', methods=['DELETE'])
 @jwt_required
 def delete_a_product(product_id):
+    if user_check() is False:
+        return jsonify({"Alert": "You're not Authorized to perform action"}), 401
+
+    if not product_cls.get_one_product_by_id(product_id):
+        return jsonify({"Not Found":
+                        "The product with ID '{0}' was not found".format(product_id)}), 404
     product_cls.delete_a_product(product_id)
-    return jsonify({"Deleted":"Product was deleted successfully"}), 200
+    return jsonify({"Deleted": "Product was deleted successfully"}), 200
+
+@prodbp.route('admin/products/category', methods=['POST'])
+@jwt_required
+def create_new_category():
+    data = request.json
+    category_name = data['category_name']
+
+    if user_check() is False:
+        return jsonify({"Alert": "You're not Authorized to perform action"}), 401
+
+    if product_cls.get_products_by_category(category_name):
+        return jsonify({
+            "Message":"Category with name '{0}' already exist".format(category_name)})
+
+    product_cls.create_category(category_name)
+    return jsonify({"Created":"The '{0}' category has been created"})
 
 @prodbp.route('admin/products/<product_id>', methods=['PUT'])
 @jwt_required
 def edit_product(product_id):
     data = request.json
 
+    product_name = data['product_name']
     product_stock = int(data['product_stock'])
     product_price = int(data['product_price'])
-    product_cls.edit_a_product(product_id, product_stock, product_price)
+
+    if user_check() is False:
+        return jsonify({"Alert": "You're not Authorized to perform action"}), 401
+
+    if not product_cls.get_one_product_by_id(product_id):
+        return jsonify({"Not Found":
+                        "The product with ID '{0}' was not found".format(product_id)}), 404
+
+    product_cls.edit_a_product(
+        product_id, product_name, product_stock, product_price)
 
     return jsonify({"Updated":
                     "Product was updated successfully"}), 200
